Extract events from order information
