+++ src/main/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringdatastoreapi/client/EmDatastoreClient.kt
@@ -20,7 +20,7 @@ import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.athen
 
 // We will instantiate as new for now
 @Component
-@Profile("dev", "test", "production")
+@Profile("!integration & !mocking")
 class EmDatastoreClient : EmDatastoreClientInterface {
   private val outputBucket: String = "s3://emds-dev-athena-query-results-20240917144028307600000004"
   private val sleepLength: Long = 1000
