+++ src/main/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringdatastoreapi/service/OrderService.kt
@@ -3,11 +3,8 @@ package uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.service
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.client.AthenaRole
-import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.ContactEventList
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.DocumentList
-import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.IncidentEventList
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.KeyOrderInformation
-import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.MonitoringEventList
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.OrderInformation
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.OrderSearchCriteria
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.OrderSearchResult
@@ -65,22 +62,5 @@ class OrderService(
       documents = emptyDocumentList,
     )
   }
-
-  fun getMonitoringEvents(orderId: String, role: AthenaRole): MonitoringEventList {
-    val result = orderInformationRepository.getMonitoringEventsList(orderId, role)
-
-    return MonitoringEventList(result)
-  }
-
-  fun getIncidentEvents(orderId: String, role: AthenaRole): IncidentEventList {
-    val result = orderInformationRepository.getIncidentEventsList(orderId, role)
-
-    return IncidentEventList(result)
-  }
-
-  fun getContactEvents(orderId: String, role: AthenaRole): ContactEventList {
-    val result = orderInformationRepository.getContactEventsList(orderId, role)
-
-    return ContactEventList(result)
-  }
 }
+
