+++ src/test/kotlin/uk/gov/justice/digital/hmpps/electronicmonitoringdatastoreapi/resources/OrderControllerTest.kt
@@ -13,22 +13,14 @@ import org.springframework.http.HttpStatus
 import org.springframework.security.core.Authentication
 import org.springframework.test.context.ActiveProfiles
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.client.AthenaRole
-import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.ContactEventDetails
-import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.ContactEventList
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.DocumentList
-import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.Event
-import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.IncidentEventDetails
-import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.IncidentEventList
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.KeyOrderInformation
-import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.MonitoringEventDetails
-import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.MonitoringEventList
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.OrderInformation
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.model.SubjectHistoryReport
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.resource.OrderController
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.service.AthenaRoleService
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.service.OrderService
 import uk.gov.justice.digital.hmpps.electronicmonitoringdatastoreapi.service.internal.AuditService
-import java.time.LocalDateTime
 
 @ActiveProfiles("test")
 @JsonTest
@@ -70,7 +62,12 @@ class OrderControllerTest {
           orderStartDate = "01-02-2012",
           orderEndDate = "03-04-2013",
         ),
-        subjectHistoryReport = SubjectHistoryReport(reportUrl = "#", name = "1234567", createdOn = "01-02-2020", time = "0900"),
+        subjectHistoryReport = SubjectHistoryReport(
+          reportUrl = "#",
+          name = "1234567",
+          createdOn = "01-02-2020",
+          time = "0900"
+        ),
         documents = DocumentList(
           pageSize = 0,
           orderDocuments = listOf(),
@@ -108,7 +105,12 @@ class OrderControllerTest {
           orderStartDate = "01-02-2012",
           orderEndDate = "03-04-2013",
         ),
-        subjectHistoryReport = SubjectHistoryReport(reportUrl = "#", name = "1234567", createdOn = "01-02-2020", time = "0900"),
+        subjectHistoryReport = SubjectHistoryReport(
+          reportUrl = "#",
+          name = "1234567",
+          createdOn = "01-02-2020",
+          time = "0900"
+        ),
         documents = DocumentList(
           pageSize = 0,
           orderDocuments = listOf(),
@@ -123,105 +125,6 @@ class OrderControllerTest {
       Assertions.assertThat(result.body).isEqualTo(expectedResult)
 
       Mockito.verify(orderService, times(1)).getOrderInformation(orderId, AthenaRole.DEV)
-    }
-  }
-
-  @Nested
-  inner class GetMonitoringEvents {
-    @Test
-    fun `gets order information from order service`() {
-      val orderId = "1ab"
-      val expectedResult = MonitoringEventList(
-        pageSize = 1,
-        events = listOf<Event<MonitoringEventDetails>>(
-          Event<MonitoringEventDetails>(
-            legacyOrderId = 123,
-            legacySubjectId = 1543,
-            type = "TEST_STATUS",
-            dateTime = LocalDateTime.of(2021, 1, 1, 1, 1, 1),
-            details = MonitoringEventDetails(
-              processedDateTime = LocalDateTime.of(2021, 1, 1, 1, 1, 2),
-            ),
-          ),
-        ),
-      )
-
-      `when`(orderService.getMonitoringEvents(orderId, AthenaRole.DEV)).thenReturn(expectedResult)
-
-      val result = controller.getMonitoringEvents(authentication, orderId)
-
-      Assertions.assertThat(result.statusCode).isEqualTo(HttpStatus.OK)
-      Assertions.assertThat(result.body).isEqualTo(expectedResult)
-
-      Mockito.verify(orderService, times(1)).getMonitoringEvents(orderId, AthenaRole.DEV)
-    }
-  }
-
-  @Nested
-  inner class GetViolationAlerts {
-    @Test
-    fun `gets order information from order service`() {
-      val orderId = "1ab"
-      val expectedResult = IncidentEventList(
-        pageSize = 1,
-        events = listOf<Event<IncidentEventDetails>>(
-          Event<IncidentEventDetails>(
-            legacyOrderId = 123,
-            legacySubjectId = 1543,
-            type = "TEST_STATUS",
-            dateTime = LocalDateTime.of(2021, 1, 1, 1, 1, 1),
-            details = IncidentEventDetails(
-              violation = "TEST_VIOLATION",
-            ),
-          ),
-        ),
-      )
-
-      `when`(orderService.getIncidentEvents(orderId, AthenaRole.DEV)).thenReturn(expectedResult)
-
-      val result = controller.getViolationAlerts(authentication, orderId)
-
-      Assertions.assertThat(result.statusCode).isEqualTo(HttpStatus.OK)
-      Assertions.assertThat(result.body).isEqualTo(expectedResult)
-
-      Mockito.verify(orderService, times(1)).getIncidentEvents(orderId, AthenaRole.DEV)
-    }
-  }
-
-  @Nested
-  inner class GetContactEvents {
-    @Test
-    fun `gets order information from order service`() {
-      val orderId = "1ab"
-      val expectedResult = ContactEventList(
-        pageSize = 1,
-        events = listOf<Event<ContactEventDetails>>(
-          Event<ContactEventDetails>(
-            legacyOrderId = 123,
-            legacySubjectId = 1543,
-            type = "TEST_STATUS",
-            dateTime = LocalDateTime.of(2021, 1, 1, 1, 1, 1),
-            details = ContactEventDetails(
-              outcome = null,
-              contactType = "PHONE_CALL",
-              reason = "TEST_REASON",
-              channel = "TEST_CHANNEL",
-              userId = null,
-              userName = null,
-              modifiedDateTime = LocalDateTime.of(2019, 9, 1, 12, 15, 9),
-            ),
-          ),
-        ),
-      )
-
-      `when`(orderService.getContactEvents(orderId, AthenaRole.DEV)).thenReturn(expectedResult)
-
-      val result = controller.getContactEvents(authentication, orderId)
-
-      Assertions.assertThat(result.statusCode).isEqualTo(HttpStatus.OK)
-      Assertions.assertThat(result.body).isEqualTo(expectedResult)
-
-      Mockito.verify(orderService, times(1)).getContactEvents(orderId, AthenaRole.DEV)
     }
   }
 }
